# bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "AssetManagementStoreApp, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/AssetInfo/GetAssets": {
      "get": {
        "tags": [
          "AssetInfo"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/AssetInfo"
              }
            }
          }
        }
      }
    },
    "/AssetInfo/GetAssetInfo/{assetId}": {
      "get": {
        "tags": [
          "AssetInfo"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "assetId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/AssetInfo/CreateAndEditAsset": {
      "post": {
        "tags": [
          "AssetInfo"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AssetInfo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StringTuple"
            }
          }
        }
      }
    },
    "/AssetInfo/DeleteAsset": {
      "delete": {
        "tags": [
          "AssetInfo"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "assetId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StringTuple"
            }
          }
        }
      }
    },
    "/Categories/GetCategories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Category"
              }
            }
          }
        }
      }
    },
    "/FileInfoAndUploads/GetFileInfo/{fileId}": {
      "get": {
        "tags": [
          "FileInfoAndUploads"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/FileInfoAndUploads/upload": {
      "post": {
        "tags": [
          "FileInfoAndUploads"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "file",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FileData"
            }
          }
        }
      }
    },
    "/FileInfoAndUploads/download/{id}": {
      "get": {
        "tags": [
          "FileInfoAndUploads"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/FileInfoAndUploads/downloadMultipleFiles/{id}": {
      "get": {
        "tags": [
          "FileInfoAndUploads"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/FileInfoAndUploads/EditFile": {
      "post": {
        "tags": [
          "FileInfoAndUploads"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "file",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/FileData"
            }
          }
        }
      }
    },
    "/FileInfoAndUploads/DeleteFile": {
      "delete": {
        "tags": [
          "FileInfoAndUploads"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Idea/GetIdeas": {
      "get": {
        "tags": [
          "Idea"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Idea"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "AssetInfo": {
      "required": [
        "comments",
        "description",
        "uploadedBy",
        "uploadedOn"
      ],
      "type": "object",
      "properties": {
        "assetId": {
          "format": "int32",
          "type": "integer"
        },
        "technologyUsed": {
          "maxLength": 500,
          "type": "string"
        },
        "uploadedBy": {
          "maxLength": 100,
          "type": "string"
        },
        "description": {
          "maxLength": 100000,
          "type": "string"
        },
        "uploadedOn": {
          "format": "date-time",
          "type": "string"
        },
        "comments": {
          "maxLength": 100000,
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        },
        "categoryId": {
          "format": "int32",
          "type": "integer"
        },
        "ideaId": {
          "format": "int32",
          "type": "integer"
        },
        "categoryName": {
          "type": "string"
        },
        "ideaName": {
          "type": "string"
        },
        "objIdea": {
          "$ref": "#/definitions/Idea"
        },
        "files": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/FileData"
          }
        },
        "categories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Category"
          }
        }
      }
    },
    "Category": {
      "required": [
        "categoryName"
      ],
      "type": "object",
      "properties": {
        "categoryId": {
          "format": "int32",
          "type": "integer"
        },
        "categoryName": {
          "maxLength": 100,
          "type": "string"
        },
        "ideas": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Idea"
          }
        }
      }
    },
    "FileData": {
      "required": [
        "dateTime",
        "fileContent",
        "fileName",
        "fileType"
      ],
      "type": "object",
      "properties": {
        "fileId": {
          "format": "int32",
          "type": "integer"
        },
        "fileName": {
          "maxLength": 200,
          "type": "string"
        },
        "fileContent": {
          "format": "byte",
          "type": "string"
        },
        "fileType": {
          "maxLength": 200,
          "type": "string"
        },
        "dateTime": {
          "format": "date-time",
          "type": "string"
        },
        "isUploaded": {
          "type": "boolean"
        },
        "assetId": {
          "format": "int32",
          "type": "integer"
        },
        "isReplaced": {
          "type": "boolean"
        }
      }
    },
    "Idea": {
      "required": [
        "description"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 200,
          "type": "string"
        },
        "description": {
          "maxLength": 100000,
          "type": "string"
        }
      }
    },
    "StringTuple": {
      "type": "object",
      "properties": {
        "item1": {
          "type": "string"
        }
      }
    }
  }
}